fix: #61 빼먹은 코드때문에 런타임 에러나는 부분들 수정

+++ data/src/main/java/com/mashup/alcoholfree/data/dto/TierResponse.kt
@@ -12,7 +12,7 @@ internal data class TierResponse(
     @SerializedName("subTitleText") val subTitleText: String,
     @SerializedName("badgeImageUrl") val badgeImageUrl: String,
     @SerializedName("cardImageUrl") val cardImageUrl: String,
-    @SerializedName("alcoholAmount") val alcoholAmount: Int,
+    @SerializedName("alcoholAmount") val alcoholAmount: Float,
 ) {
     fun toDomainModel(): AlcoholTier {
         return AlcoholTier(
